[NativeCompanion] Refactor widget data fetching code
